finish building API

+++ backend/controllers/product.controller.js
@@ -46,11 +46,17 @@ export const updateProduct = async (req, res) =>{
 }
 export const deleteProduct = async (req, res ) =>{
     const {id} = req.params
+
+    if(!product.name || !product.price || !product.image){
+        return res.status(400).json({success: false, message: "Please provide all fields"});
+    }
+
+    
     try {
         await Product.findByIdAndDelete(id);
         res.status(200).json({success: true, message: "Product deleted"});
     } catch (error) {
         console.log("error in deleting product: ", error.message);
-        res.status(404).json({success:false, message: "Product not found"});
+        res.status(500).json({success:false, message: "Sever Error"});
     }
 }
